[Auto Parallel] Enable to_static unittest
